add firebase login to login component
