+++ lib/app_component.html
@@ -1,22 +1,1 @@
 <router-outlet [routes]="Routes.all"></router-outlet>
-<div *ngIf="!isAuthenticated()">
-    <p>Login</p>
-</div>
-
-<firebase-auth-ui [disableAutoSignIn]="false" [uiConfig]="getUIConfig()"></firebase-auth-ui>
-
-<br/>
-<div *ngIf="isAuthenticated()">
-  <button (click)="logout()">Logout</button>
-
-  Authenticated!
-  <p>User email: {{userEmail}}  Display Name: {{displayName}}</p>
-  <p>User Json: {{userJson}}</p>
-  <p>Provider Access Token 
-  (may not be present unless session is new) : {{providerAccessToken}}</p>
-
-  
-  <todo-list></todo-list>
-  <today-list></today-list>
-  
-</div>
