+++ web/index.html
@@ -20,7 +20,6 @@
     <script src="https://cdn.firebase.com/libs/firebaseui/3.4.1/firebaseui.js"></script>
     <link type="text/css" rel="stylesheet" href="https://cdn.firebase.com/libs/firebaseui/3.4.1/firebaseui.css" />
   </head>
->>>>>>> add basic firebase auth features
   <body>
     <my-app>Loading...</my-app>
   </body>
